+++ src/Sulu/Component/Content/Tests/Unit/Compat/Structure/StructureBridgeTest.php
@@ -166,4 +166,21 @@ class StructureBridgeTest extends \PHPUnit_Framework_TestCase
         $this->assertFalse($structure->getIsShadow());
         $this->assertNull($structure->getShadowBaseLanguage());
     }
+
+    public function testWithoutDocument()
+    {
+        $metadata = $this->prophesize(StructureMetadata::class);
+        $inspector = $this->prophesize(DocumentInspector::class);
+        $propertyFactory = $this->prophesize(LegacyPropertyFactory::class);
+
+        $structure = new StructureBridge(
+            $metadata->reveal(),
+            $inspector->reveal(),
+            $propertyFactory->reveal(),
+            null
+        );
+
+        $this->assertNull($structure->getWebspaceKey());
+        $this->assertFalse($structure->getIsShadow());
+    }
 }