+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG for Sulu
 ==================
 
+* 1.6.5 (2017-10-04)
+    * HOTFIX      #---- [Husky]                   Fixed bug in `escapeHtml` method
+    * HOTFIX      #3535 [Content]                 Fix bug in structure bridge when no document is available
+    * HOTFIX      #3514 [ContentBundle]           Fixed override route when using "tree_full_edit"
+    * HOTFIX      #3541 [CategoryBundle]          Fixed category-list for null selected value
+    * HOTFIX      #3543 [PreviewBundle]           Fixed rerender preview if html is the same
+    * HOTFIX      #3543 [ContentBundle]           Fixed update resource-locator with date content type
+    * ENHANCEMENT #3544 [SnippetBundle]           Added reference-store to snippet data-provider
+
 * 1.6.4 (2017-09-14)
     * HOTFIX      #3509 [MediaBundle]             Fixed deletion of tag when referenced in media fileversion
     * HOTFIX      #3504 [SnippetBundle]           Fixed snippet areas with uppercases